+++ src/redux/documentation/documentationSlice.ts
@@ -1,15 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
-export interface AssetMetaData {
-  author: string;
-  description: string;
-  identifiers?: string[];
-  license: string;
-  name: string;
-  path: string;
-  url: string;
-  version: string;
-}
+import { AssetMetaData } from '@/types/types';
 
 export interface DocumentationState {
   isInitialized: boolean;
